requirements.txt
Flask==2.3.3
stripe==5.5.0
python-dotenv==1.0.0